WIP: Port klipper to use wayland clipboard
Summary:
Wayland has an entire new protocol for getting and setting clipboard
when we don't have focus. Unfortunately this means reinventing
QClipboard from the QPA.

Not mergable as-is, especially the hardcoded line in CMakeLists but
uploaded to show direction.

It uses the newly agreed approach of using QtWayland generated classes
in clients directly rather than writing full pimpl API-fixed wrappers
first,
as ultimately that didn't really help do anything.

Code is written so that it can be moved to KWindowSystem if needed by
KDEConnect or others.

Test Plan: Copying and pasting all over the place

Reviewers: #kwin

Subscribers: ngraham, cblack, apol, nicolasfella, zzag, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29476